requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
pywin32==311
python-dotenv==1.0.0
bcrypt==4.1.1
PyJWT==2.12.1
reportlab==4.0.9
openpyxl==3.1.2
requests==2.31.0
msal==1.26.0
